[STYLE] clean codes
