+++ bentoml/adapters/tensorflow_tensor_output.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import base64
 import json
 import itertools
 from typing import Iterable
@@ -21,13 +20,11 @@ import numpy as np
 import argparse
 
 from bentoml.marshal.utils import SimpleResponse, SimpleRequest
+from bentoml.adapters.utils import TfTensorJsonEncoder
 from bentoml.adapters.base_output import BaseOutputAdapter
 
 
-TF_B64_KEY = "b64"
-
-
-def to_numpy(tensor):
+def tf_to_numpy(tensor):
     '''
     Tensor -> ndarray
     List[Tensor] -> tuple[ndarray]
@@ -35,47 +32,13 @@ def to_numpy(tensor):
     import tensorflow as tf
 
     if isinstance(tensor, (list, tuple)):
-        return tuple(to_numpy(t) for t in tensor)
+        return tuple(tf_to_numpy(t) for t in tensor)
 
     if tf.__version__.startswith("1."):
         with tf.compat.v1.Session():
             return tensor.numpy()
     else:
         return tensor.numpy()
-
-
-class TfTensorJsonEncoder(json.JSONEncoder):
-    """ Special json encoder for numpy types """
-
-    def default(self, o):  # pylint: disable=method-hidden
-
-        import tensorflow as tf
-
-        # Tensor -> ndarray or object
-        if isinstance(o, tf.Tensor):
-            if tf.__version__.startswith("1."):
-                with tf.compat.v1.Session():
-                    return o.numpy()
-            else:
-                return o.numpy()
-        if isinstance(o, np.generic):
-            return o.item()
-        if isinstance(o, np.ndarray):
-            return o.tolist()
-        if isinstance(o, bytes):
-            try:
-                return o.decode('utf-8')
-            except UnicodeDecodeError:
-                return {TF_B64_KEY: base64.b64encode(o).decode("utf-8")}
-        return json.JSONEncoder.default(self, o)
-
-
-def jsonize(result, indent=None):
-    try:
-        return json.dumps(result, cls=TfTensorJsonEncoder, indent=indent)
-    except (TypeError, OverflowError):
-        # when result is not JSON serializable
-        return json.dumps({"result": str(result)})
 
 
 class TfTensorOutput(BaseOutputAdapter):
@@ -99,7 +62,7 @@ class TfTensorOutput(BaseOutputAdapter):
         requests: Iterable[SimpleRequest] = None,
     ) -> Iterable[SimpleResponse]:
         # TODO(bojiang): header content_type
-        results = to_numpy(result_conc)
+        results = tf_to_numpy(result_conc)
         assert isinstance(results, np.ndarray)
         if slices is None:
             slices = [i for i in range(results.shape[0])]
@@ -107,16 +70,12 @@ class TfTensorOutput(BaseOutputAdapter):
             fallbacks = itertools.repeat(None)
         responses = [None] * len(slices)
 
-        for i, (s, f, r) in enumerate(zip(slices, fallbacks, requests)):
+        for i, (s, f) in enumerate(zip(slices, fallbacks)):
             if s is None:
                 responses[i] = f
                 continue
-            batch_flag = self.is_batch_request(r)
             result = results[s]
-            if batch_flag:
-                result_str = jsonize(result[0])
-            else:
-                result_str = jsonize(result)
+            result_str = json.dumps(result, cls=TfTensorJsonEncoder)
             responses[i] = SimpleResponse(200, dict(), result_str)
         return responses
 
@@ -132,13 +91,13 @@ class TfTensorOutput(BaseOutputAdapter):
         parser.add_argument("-o", "--output", default="json", choices=["str", "json"])
         parsed_args = parser.parse_args(args)
         if parsed_args.output == 'json':
-            result = jsonize(result, indent='  ')
+            result = json.dumps(result, cls=TfTensorJsonEncoder, indent='  ')
         else:
             result = str(result)
         print(result)
 
     def to_aws_lambda_event(self, result, event):
-        result = jsonize(result)
+        result = json.dumps(result, cls=TfTensorJsonEncoder)
         # Allow disabling CORS by setting it to None
         if self.cors:
             return {
