+++ bentoml/adapters/utils.py
@@ -1,3 +1,59 @@
+import json
+
+
+TF_B64_KEY = "b64"
+
+
+class B64JsonEncoder(json.JSONEncoder):
+    """ Special json encoder for numpy types """
+
+    def default(self, o):  # pylint: disable=method-hidden
+        import base64
+
+        if isinstance(o, bytes):
+            try:
+                return o.decode('utf-8')
+            except UnicodeDecodeError:
+                return {TF_B64_KEY: base64.b64encode(o).decode("utf-8")}
+
+        try:
+            return super(B64JsonEncoder, self).default(o)
+        except (TypeError, OverflowError):
+            return {"unknown_obj": str(o)}
+
+
+class NumpyJsonEncoder(B64JsonEncoder):
+    """ Special json encoder for numpy types """
+
+    def default(self, o):  # pylint: disable=method-hidden
+        import numpy as np
+
+        if isinstance(o, np.generic):
+            return o.item()
+
+        if isinstance(o, np.ndarray):
+            return o.tolist()
+
+        return super(NumpyJsonEncoder, self).default(o)
+
+
+class TfTensorJsonEncoder(NumpyJsonEncoder):
+    """ Special json encoder for numpy types """
+
+    def default(self, o):  # pylint: disable=method-hidden
+
+        import tensorflow as tf
+
+        # Tensor -> ndarray or object
+        if isinstance(o, tf.Tensor):
+            if tf.__version__.startswith("1."):
+                with tf.compat.v1.Session():
+                    return o.numpy()
+            else:
+                return o.numpy()
+        return super(TfTensorJsonEncoder, self).default(o)
+
+
 def concat_list(lst, batch_flags=None):
     """
     >>> lst = [
@@ -24,5 +80,6 @@ def concat_list(lst, batch_flags=None):
         else:
             datas.append(r)
             slices[i] = row_flag
+            j += 1
         row_flag += j + 1
     return datas, slices
